+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 <cite>Release contributors</cite>
 
 - 6 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.3.2+author%3Amlytvyn+is%3Apr)
-- 1 PR(s) by [Mihai Sprinceana](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.3.2+author%3AMihaiSprinceana+is%3Apr)
+- 2 PR(s) by [Mihai Sprinceana](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.3.2+author%3AMihaiSprinceana+is%3Apr)
 
 ### `Project Import` enhancements
 - Update language level in BGT write thread [#1719](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1719)
@@ -12,6 +12,7 @@
 - Show detailed progress of the decompilation progress [#1727](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1727)
 - Show all detected modules if `platform` is not found [#1725](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1725)
 - Preselect only first conflicting extension according to load order [#1726](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1726)
+- Excludes decompiled sources output (doc/decompiledsrc) from module content roots. [#1727](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1727)
 
 ### `WSL` enhancements
 - Support resolution of the relative symbolic links on WSL [#1724](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1724)